Add golden model tests (15 tests) and fix 6 bugs
## Golden model test infrastructure
- tests/asm/*.s: 15 test source files covering all instruction groups
- tests/golden/*.o: reference output generated by original HAS060.X v3.09+91
- tests/golden_test.rs: test runner that assembles each source with rhas and
  compares byte-for-byte against golden reference
- tests/gen_golden.sh: script to regenerate golden files using run68 + HAS060.X

## Bug fixes (all discovered via golden test failures)

### pass2.rs: preserve explicit .w branch size
- Only optimize branches with no explicit size specifier (req_size = None)
- Previously, explicit `.w` branches were incorrectly optimized to `.s`

### pass3.rs: use $3000 reserve record for .ds in all sections
- `.ds.X` in text/data sections now emits `$3000` record (HAS-compatible)
  instead of actual zero bytes
- Added emit_reserve() method; flush dsb_pending before emitting in non-BSS

### pass1.rs: .elseif/.else skip handling fix
- Added if_matched[65] to track matched branches at each nesting level
- Prevents .else from executing after a .elseif branch already matched

### pass1.rs: resolve .set symbols immediately in pass1
- When encode_insn fails with DeferToLinker but resolved ops succeed,
  emit as Const immediately instead of deferring to pass3
- Fixes .set variables being evaluated with their final value in pass3
  instead of the value at the point of use

### pass1.rs + addressing/mod.rs: .reg register list support in MOVEM
- Added parse_reg_list_mask() to addressing/mod.rs
- .reg handler now uses register list parser to store computed mask as
  ValueWord constant in RegSym.define
- parse_ea() recognizes RegSym and returns Immediate(mask) for MOVEM use

### instructions/mod.rs + expr/eval.rs: earlier bug fixes
- MOVEP: fix base opcode 0x0100 -> 0x0108
- ValueWord: remove incorrect sign extension in eval_rpn

+++ src/expr/eval.rs
@@ -74,7 +74,7 @@ pub fn eval_rpn(
 
             // ---- 即値 ----
             RPNToken::ValueByte(v) => stack.push(EvalValue::constant(*v as i32)),
-            RPNToken::ValueWord(v) => stack.push(EvalValue::constant(*v as i16 as i32)),
+            RPNToken::ValueWord(v) => stack.push(EvalValue::constant(*v as i32)),
             RPNToken::Value(v)     => stack.push(EvalValue::constant(*v as i32)),
 
             // ---- シンボル参照 ----
